# Load models much faster on accelerator devices!!