--- a/src/transformers/modeling_utils.py
+++ b/src/transformers/modeling_utils.py
@@ -4838,7 +4838,7 @@ class PreTrainedModel(nn.Module, ModuleUtilsMixin, GenerationMixin, PushToHubMix
             offload_index = {
                 p[len(start_prefix) :]: {"safetensors_file": f, "weight_name": p, "dtype": str_dtype}
                 for p, f in weight_map.items()
-                if p.startswith(start_prefix) and expanded_device_map[p[len(start_prefix) :]] == "disk"
+                if p.startswith(start_prefix) and param_device_map[p[len(start_prefix) :]] == "disk"
             }
         else:
             offload_index = None